smart_agent/requirements.txt
uvicorn==0.24.0
requests==2.31.0
fastapi==0.104.1
anthropic==0.39.0
pydantic==1.10.10
python-dotenv==1.0.0
validators==0.20.0
PyYAML==6.0.1
mangum==0.17.0
boto3==1.34.0
botocore==1.34.0
